Add files via upload
1) 
Changed layer initialization so that input example shape is passed instead of solely the feature axis dimension. The output shape is now an attribute of each layer, but those that always do not change the input shape have is set as None.

2)
Now not every layer has a custom initialization method. Only those that change the data shape through propagation or have trainable parameters that need initialization have the method.